+++ BackEnd/Administrator/src/main/java/com/football/administrator/Service/Impl/AnnouncementServiceImpl.java
@@ -27,7 +27,7 @@ public class AnnouncementServiceImpl implements AnnouncementService {
     }
 
     @Override
-    public List<AnnouncementDTO> getAnnouncementList(Integer id) {
+    public List<AnnouncementDTO> getAnnouncementList(Long id) {
         return announcementmapper.getAnnouncementList(id);
     }
 }
